added first draft of PTL implementation (UNTESTED!)
first of all, I didn't test this yet. I didn't even run it yet. The potential optimization of having a register_buffer  instead of a vanilla python list is missing. (though I'm not sure if that will actually change anything, as the data in there should in theory be replicated across all gpus because it's part of the model; also, if we use register_buffers we might need to make sure that only the first gpu-process adds to the queue because we might add the same stuff multiple times to the queue otherwise)