ported my new_ds PT model to PTL
this is mostly untested, but I attempted to get rid of most bugs on a windows machine that had too little gpu memory to run the full training set. But the sanity test for the validation set passed, so I think things are looking decent. Next, I'll have to use ddp and deepspeed and test this on the server.